Quiet ThreadSanitizer
This is a false-positive, but only because programs semantics meant that
accessing the queue.count without taking the lock was "ok".  However,
tsan doesn't know this.  By moving the queue.count=0 line to before the
associated mutex is created, tsan doesn't warn any more.

Note that neither worker->exit nor worker->queue.count need to be set to
0, as that's done by the bzero call.