silence trivial race complaints
We have a race when we calculate the mean time to send an event.

This doesn't matter much, as it is just a reference stat. Which
is why we dont use locking.

Anyway, thread-santitizer complains, so this silences the complaints.

Also we check both elapsed_use and total, which is safer. We don't
want a divide by zero here.

The race is still there, but thread-sanitizer wont report it, and
this is better code anyway.

Thanks to Andrea Guzzo for finding this.