graphitize jemalloc stats if available.
Unfortunately it seems that with the CentOS jemalloc
the stats are not available.  jemalloc needs to be configured
with --enable-stats, and then mallctl config.stats would return true.
It returns false.

The tricky part about this patch is using dlopen/dlsym so that
the code builds and runs even without the jemalloc.so.  The jemalloc
header needs to be present, though, so the jemalloc-devel pkg needs
to be installed.